# Reduced code size and optimized name key creation for JSON properties and parameters