@Tornhoof 2020-07-30
Are there any statistics on average json property length? It might be worth to calculate that e.g. using https://www.schemastore.org/json/ to figure that out. There perf regressions for even length values up to 6 byte length might warrant that.