@YohDeadfall 2020-07-30
It really depends on a user's choice and preferences. For example, an SKU identifier can be named three different ways without taking casing into account: `sku`, `sku_id` or `SkuId`.

Speaking about the regression, it's caused by the JIT. For some reason it puts `rcx` at `rsp` before checking length's parity. If it's even then it reads `rcx` back from memory located at `rsp`. @EgorBo, could you explain this weird behavior?

### Before fix

```assembly
; Expirements.Demo.Program.<GetKeyOptimized>g__GetKey|1_0(System.ReadOnlySpan`1<Byte>)
       push      rax
       mov       rax,[rcx]
       mov       edx,[rcx+8]
       cmp       edx,7
       jle       short M01_L02
       mov       rcx,0FFFFFFFFFFFF
       and       rcx,[rax]
       cmp       edx,0FF
       jle       short M01_L00
       mov       eax,0FF
       jmp       short M01_L01
M01_L00:
       mov       eax,edx
M01_L01:
       mov       edx,eax
       jmp       short M01_L07
M01_L02:
       cmp       edx,5
       jg        short M01_L05
       cmp       edx,3
       jg        short M01_L04
       cmp       edx,1
       jg        short M01_L03
       xor       ecx,ecx
       jmp       short M01_L06
M01_L03:
       movzx     ecx,word ptr [rax]
       jmp       short M01_L06
M01_L04:
       mov       ecx,[rax]
       jmp       short M01_L06
M01_L05:
       mov       ecx,[rax]
       movzx     r8d,word ptr [rax+4]
       shl       r8,20
       or        rcx,r8
M01_L06:
       mov       [rsp],rcx
       test      dl,1
       je        short M01_L08
       lea       r8d,[rdx+0FFFF]
       movsxd    r9,r8d
       movzx     eax,byte ptr [rax+r9]
       mov       ecx,r8d
       shl       ecx,3
       shl       rax,cl
       mov       rcx,rax
       or        rcx,[rsp]
M01_L07:
       movsxd    rax,edx
       shl       rax,38
       or        rcx,rax
       mov       rax,rcx
       add       rsp,8
       ret
M01_L08:
       mov       rcx,[rsp]
       jmp       short M01_L07
; Total bytes of code 149
```

### After fix

```assembly
; Expirements.Demo.Program.<GetKeyOptimized>g__GetKey|2_0(System.ReadOnlySpan`1<Byte>)
       mov       rax,[rcx]
       mov       ecx,[rcx+8]
       cmp       ecx,7
       jle       short M01_L02
       mov       rdx,0FFFFFFFFFFFF
       and       rdx,[rax]
       cmp       ecx,0FF
       jle       short M01_L00
       mov       r8d,0FF
       jmp       short M01_L01
M01_L00:
       mov       r8d,ecx
M01_L01:
       mov       ecx,r8d
       jmp       short M01_L06
M01_L02:
       cmp       ecx,5
       jg        short M01_L05
       cmp       ecx,3
       jg        short M01_L04
       cmp       ecx,1
       jg        short M01_L03
       xor       edx,edx
       jmp       short M01_L06
M01_L03:
       movzx     edx,word ptr [rax]
       jmp       short M01_L06
M01_L04:
       mov       edx,[rax]
       jmp       short M01_L06
M01_L05:
       mov       edx,[rax]
       movzx     r8d,word ptr [rax+4]
       shl       r8,20
       or        rdx,r8
M01_L06:
       movsxd    r8,ecx
       shl       r8,38
       or        rdx,r8
       test      cl,1
       je        short M01_L07
       dec       ecx
       movsxd    r8,ecx
       movzx     eax,byte ptr [rax+r8]
       shl       ecx,3
       shl       rax,cl
       or        rax,rdx
       mov       rdx,rax
M01_L07:
       mov       rax,rdx
       ret
; Total bytes of code 131
```