@YohDeadfall 2020-07-30
Results after the code reordering:

|             Method |   value |     Mean |     Error |    StdDev | Code Size |
|------------------- |-------- |---------:|----------:|----------:|----------:|
|    **GetKeyOptimized** | **Byte[0]** | **4.907 ns** | **0.1144 ns** | **0.1488 ns** |     **187 B** |
|      GetKeyRuntime | Byte[0] | 5.278 ns | 0.0128 ns | 0.0100 ns |     422 B |
|    **GetKeyOptimized** | **Byte[1]** | **4.437 ns** | **0.0629 ns** | **0.0557 ns** |     **187 B** |
|      GetKeyRuntime | Byte[1] | 5.439 ns | 0.0718 ns | 0.0636 ns |     422 B |
|    **GetKeyOptimized** | **Byte[2]** | **4.662 ns** | **0.1287 ns** | **0.1804 ns** |     **187 B** |
|      GetKeyRuntime | Byte[2] | 5.382 ns | 0.1440 ns | 0.2326 ns |     422 B |
|    **GetKeyOptimized** | **Byte[3]** | **4.517 ns** | **0.1259 ns** | **0.1499 ns** |     **187 B** |
|      GetKeyRuntime | Byte[3] | 4.781 ns | 0.0127 ns | 0.0112 ns |     422 B |
|    **GetKeyOptimized** | **Byte[4]** | **4.518 ns** | **0.0271 ns** | **0.0226 ns** |     **187 B** |
|      GetKeyRuntime | Byte[4] | 5.311 ns | 0.0098 ns | 0.0087 ns |     422 B |
|    **GetKeyOptimized** | **Byte[5]** | **4.247 ns** | **0.0095 ns** | **0.0079 ns** |     **187 B** |
|      GetKeyRuntime | Byte[5] | 4.994 ns | 0.0023 ns | 0.0019 ns |     422 B |
|    **GetKeyOptimized** | **Byte[6]** | **4.553 ns** | **0.0606 ns** | **0.0567 ns** |     **187 B** |
|      GetKeyRuntime | Byte[6] | 4.736 ns | 0.0031 ns | 0.0029 ns |     422 B |
|    **GetKeyOptimized** | **Byte[7]** | **3.960 ns** | **0.0245 ns** | **0.0205 ns** |     **187 B** |
|      GetKeyRuntime | Byte[7] | 4.252 ns | 0.0028 ns | 0.0025 ns |     422 B |
|    **GetKeyOptimized** | **Byte[8]** | **4.237 ns** | **0.0134 ns** | **0.0112 ns** |     **187 B** |
|      GetKeyRuntime | Byte[8] | 4.244 ns | 0.0145 ns | 0.0121 ns |     422 B |
|    **GetKeyOptimized** | **Byte[9]** | **3.976 ns** | **0.0070 ns** | **0.0062 ns** |     **187 B** |
|      GetKeyRuntime | Byte[9] | 4.316 ns | 0.0760 ns | 0.0711 ns |     422 B |

/cc @stephentoub