@steveharter 2020-08-06
@YohDeadfall thanks for looking into this area. It is a critical hot path for deserialization.

Do you know if adding  `AggressiveOptimization` helps? I assume you run the tests enough times so the tiered JIT kicks in:
```[MethodImpl(MethodImplOptions.AggressiveOptimization| MethodImplOptions.AggressiveInlining)]```