@steveharter 2020-08-06
The noted regression then is `Byte[6]`, correct?

```
GetKeyOptimized | Byte[6] | 5.016 ns | 0.0174 ns | 0.0163 ns | 199 B
GetKeyRuntime | Byte[6] | 4.738 ns | 0.0012 ns | 0.0010 ns | 422 B
```